requirements.txt
python-telegram-bot<=19.0
psycopg2
peewee
python-dotenv
jaconv
tqdm
